No change tooltip after close tab via middle click, added snap to other points in vector visuals

+++ Kaiplayer/SpellCheckerDialog.cpp
@@ -103,7 +103,7 @@ wxString SpellCheckerDialog::FindNextMisspell()
 		lastLine = lastActiveLine = tab->Grid->currentLine;
 		lastMisspell = 0;
 	}
-	for(int i = lastLine; i < tab->Grid->GetCount(); i++){
+	for(size_t i = lastLine; i < tab->Grid->GetCount(); i++){
 		errors.clear();
 		Dialogue *Dial = tab->Grid->GetDialogue(i);
 		if(Dial->IsComment && noComments){continue;}
